Add components to the 'dbz' module.
Create a interface to define the 'Character' object structure.
Create a component that show a list of characters; the characters are passed to the component from its parent by using '@Input()'.
Create a component that show a form to add a new character; the character is passed to the parent by using '@Output()'.