Add services to the 'dbz' module.
Create the service 'dbz.service.ts' file to keep data. Move data from 'main-page-component.ts' to the service file.
Inject the service in the 'main-page-component.ts'.
Add the UUID package using 'npm 1 uuid' and 'npm i --save-dev @types/uuid'. This package is used to assign to every character a unique id.
Change the 'onDeleteCharacter' method so it uses the unique ID instead of the index.
Add the 'id' property in the interface file, so all the 'Character' objects have this new property.